refactor: controller 위치 수정
