+++ storage/src/main/java/com/medicall/storage/db/core/treatment/PrescriptionRepository.java
@@ -6,5 +6,4 @@ import java.util.List;
 public interface PrescriptionRepository extends JpaRepository<Prescription, Long> {
     @Query("SELECT p FROM Prescription p WHERE p.member.id = :memberId")
     List<Prescription> findAllByMemberId(@Param("memberId") Long memberId);
-
 }
